+++ R/sanity.R
@@ -43,6 +43,7 @@ sanitize_output <- function(output) {
         "\\usepackage[normalem]{ulem}",
         "\\usepackage{graphicx}",
         "\\UseTblrLibrary{booktabs}",
+        "\\UseTblrLibrary{siunitx}",
         "\\NewTableCommand{\\tinytableDefineColor}[3]{\\definecolor{#1}{#2}{#3}}",
         "\\newcommand{\\tinytableTabularrayUnderline}[1]{\\underline{#1}}",
         "\\newcommand{\\tinytableTabularrayStrikeout}[1]{\\sout{#1}}"