+++ R/save_tt.R
@@ -116,6 +116,7 @@ latex_standalone <- "
 \\usepackage{codehigh}
 \\usepackage[normalem]{ulem}
 \\UseTblrLibrary{booktabs}
+\\UseTblrLibrary{siunitx}
 \\newcommand{\\tinytableTabularrayUnderline}[1]{\\underline{#1}}
 \\newcommand{\\tinytableTabularrayStrikeout}[1]{\\sout{#1}}
 \\NewTableCommand{\\tinytableDefineColor}[3]{\\definecolor{#1}{#2}{#3}}